make acounts controller and midify routing
