+++ app/controllers/pages_controller.rb
@@ -1,10 +1,14 @@
 class PagesController < ApplicationController
   before_action :sign_in_required, only: [:show]
 
+
   def index
+    # ユーザ一覧画面を表示
   end
+  
 
   def show
-    @user = current_user
+    # @user = User.find_by...?
+      # ユーザのプロフィール画面を表示
   end
 end
