--- app/controllers/softdeletes_controller.rb
@@ -1,17 +1,0 @@
-
-
-class SoftdeletesController < ApplicationController
-
-  def unsubscribe
-  end
-
-  def withdrawal
-    @user = current_user
-    # is_deletedカラムをtrueに変更することにより削除フラグを立てる
-    @user.update(is_deleted: true)
-    reset_session
-    flash[:notice] = "退会処理を実行いたしました"
-    redirect_to root_path
-  end
-
-end
