+++ app/helpers/users/accounts_helper.rb
@@ -1,0 +1,2 @@
+module Users::AccountsHelper
+end
