--- app/javascript/app.js
@@ -1,9 +1,0 @@
-
-import Vue from 'vue'
-new Vue({
-  el: '#container',
-  data: {
-    age: '4.503 billion years'
-  }
-})
-
