+++ app/views/layouts/application.html.erb
@@ -8,7 +8,6 @@
         <%= csp_meta_tag %>
         <%= stylesheet_link_tag "application", "data-turbo-track": "reload" %>
         <%= javascript_importmap_tags %>
-        <%= yield(:head) %>
     </head>
 <%# ブレイクポイント対応はしていない %>
     <body class="font-body text-stone-700 bg-lime-300 relative z-0 overflow-x-hidden">
