+++ app/views/pages/show.html.erb
@@ -1,25 +1,13 @@
-<h1><%= @user.name %>さんのページです</h1>
+<h1>プロフィールページです</h1>
 
 <h3>ユーザー名</h3>
-<%= @user.name %>
 <h3>メールアドレス</h3>
-<%= @user.email %>
 <h3>プロフィール</h3>
-<%= @user.profile%>
-
+<h3>投稿記事一覧</h3>
 
 
 <p>実装予定</p>
 <ul>
   <li>プロフィール表示</li>
-  <li>自分の投稿を表示</li>
   <li>feed?</li>
 </ul>
-
-<% content_for :head do %>
-  <%= javascript_import_module_tag "app" %>
-<% end %>
-
-  <div id="container">
-    The Planet Wercury was formed {{ age }} age.
-  </div>
