+++ app/views/users/accounts/show.html.erb
@@ -1,0 +1,18 @@
+<h1><%= @user.name %>さんのページです</h1>
+
+<h3>ユーザー名</h3>
+<%= @user.name %>
+<h3>メールアドレス</h3>
+<%= @user.email %>
+<h3>プロフィール</h3>
+<%= @user.profile%>
+
+
+
+<p>実装予定</p>
+<ul>
+  <li>画像投稿機能</li>
+  <li>プロフィール画像設定機能</li>
+  <li>マイページををログイン不要のユーザに表示</li>
+  <li>フィード</li>
+</ul>
