+++ config/routes.rb
@@ -1,8 +1,9 @@
 Rails.application.routes.draw do
+
   resources :posts
 
   root 'pages#index'
-  get "/mypage" => 'pages#show', as: 'mypage'
+  # get "/profile" => 'pages#show', as: 'profile'
   get "/about" => 'pages#about', as: 'about'
   get "/contact" => 'pages#contact', as: 'contact'
   get "/privacy" => 'pages#privacy', as: 'privacy'
@@ -17,11 +18,15 @@ Rails.application.routes.draw do
       password: 'secret', confirmation: 'verification',
       registration: 'register', edit: 'edit/profile'
     }
+  devise_scope :user do
+    get "mypage", to: "users/accounts#show"
+    get '/:id/unsubscribe' => 'users/accounts#unsubscribe', as: 'unsubscribe'
+    patch '/:id/withdrawal' => 'users/accounts#withdrawal', as: 'withdrawal'
+  end
 
-
-    # 退会確認画面
-    get '/:id/unsubscribe' => 'softdeletes#unsubscribe', as: 'unsubscribe'
-    # 論理削除用のルーティング
-    patch '/:id/withdrawal' => 'softdeletes#withdrawal', as: 'withdrawal'
+  # 退会確認画面
+  # get '/:id/unsubscribe' => 'softdeletes#unsubscribe', as: 'unsubscribe'
+  # # 論理削除用のルーティング
+  # patch '/:id/withdrawal' => 'softdeletes#withdrawal', as: 'withdrawal'
   
 end
